Add annotation functionality and refactoring
