+++ kubernetesCluster/clientset.go
@@ -1,4 +1,4 @@
-package kubernetes_functions
+package kubernetesCluster
 
 import (
 	"k8s.io/client-go/kubernetes"
